feat: extract profile page logic into component
